requirements.txt
Django==5.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.0
Pillow==10.4.0
django-filter==24.3
dj-database-url==2.2.0
gunicorn==23.0.0
whitenoise==6.8.2
pyotp==2.9.0
django-ratelimit==4.1.0
python-docx==1.1.2
reportlab==4.2.2